live_set works again

+++ lib/als_delta/ als_delta_class.rb
@@ -1,5 +1,8 @@
 require 'date'
+
 class AlsDelta
+  extend Run
+
   def initialize(set_name, **options)
     @loglevel  = "-loglevel #{options[:loglevel]}"
     @loglevel += ' -stats' unless options[:loglevel] == 'quiet'
@@ -10,14 +13,16 @@ class AlsDelta
     backup_set
 
     @contents = Zlib::GzipReader.open(set_name, &:readlines)
+
+    at_exit { cleanup }
   end
 
   def backup_name
     @backup_name = if options[:keep]
                      timestamp = Time.now.getutc.strftime '%Y-%m-%d_%H:%M:%S'
-                     "#{@backup_dir}/#{backup_name_only}.#{timestamp}.als"
+                     "#{@backup_dir}/#{backup_name_only}_#{timestamp}.als"
                    else
-                     "#{@backup_dir}/#{backup_name_only}.backup.als"
+                     "#{@backup_dir}/#{backup_name_only}_backup.als"
                    end
   end
 
@@ -26,9 +31,14 @@ class AlsDelta
   end
 
   def cleanup
+    return if options[:keep]
+    return unless File.exist? @backup_name
+
+    puts "Deleting #{@backup_name}"
     File.delete @backup_name
   end
 
   def show
+    run "zdiff '#{@set_name}' '#{@backup_name}'"
   end
 end
